# api/openapi.yaml
openapi: 3.0.3
info:
  title: LazyDeck local service API
  version: "1.0.0"
  description: |
    Versioned, localhost-only HTTP+SSE API over LazyDeck's existing SteamOS
    Devkit operations (see issue #13). Engine integrations (the Godot
    plugin in issue #14, and later a Unity package) consume this contract
    instead of importing Go packages, invoking the Python bridge directly,
    or reimplementing Valve's SteamOS Devkit protocol.

    This document is the source of truth for routes and payload shapes;
    internal/server implements it. Routes not described here (e.g. a future
    /v2) are out of scope for this file.

    launch and stop are included below as routes (POST
    /v1/devices/{deviceId}/games/{gameId}/launch and .../stop), so their URL
    shape is stable, but both always return 501 with error kind
    "unsupported" in this version, because python/cli.py has no launch/stop
    primitive yet. Check GET /v1/capabilities before calling them.
servers:
  - url: http://127.0.0.1:{port}
    description: >-
      Loopback only. Port and the required bearer token are read from the
      connection file `lazydeck serve` writes on startup (see
      internal/server/connection.go), not from this document.
    variables:
      port:
        default: "0"
security:
  - bearerAuth: []
paths:
...
  /v1/devices/{deviceId}/games:
    get:
      summary: List games currently installed/deployed on a device.
      operationId: listGames
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "200":
          description: Installed games.
          content:
            application/json:
              schema:
                type: object
                properties:
                  games:
                    type: array
                    items: {}
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/DeviceError"
...
  /v1/devices/{deviceId}/logs/sync:
    post:
      summary: Pull the device's Steam logs and minidumps into a local directory, as an async job.
      operationId: syncLogs
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [directory]
              properties:
                game_id:
                  type: string
                  pattern: '^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$'
                  description: >-
                    Accepted for forward compatibility but currently
                    unused: this operation always fetches the device's
                    complete ~/.local/share/Steam/logs and minidump
                    directories regardless of this value — it cannot yet
                    filter by title. Optional; omit it.
                directory:
                  type: string
                  description: >-
                    Absolute local directory to sync logs into. A relative
                    path is rejected — see the deployments endpoint's
                    `directory` field for why.
      responses:
        "202":
          description: Log sync accepted and queued.
          content:
            application/json:
              schema:
                type: object
                properties:
                  job:
                    $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/InvalidInput"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/DeviceBusy"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Per-launch token from the connection file `lazydeck serve` writes
        on startup. Required on every /v1/ route, including /v1/health.
  parameters:
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema: { type: string }
      description: A device's configured name (devices.toml [[device]].name).
...
  schemas:
...
    ApiError:
      type: object
      properties:
        kind:
          type: string
          description: >-
            Stable machine-readable reason. Known values: unauthorized,
            invalid-input, not-found, device-busy, unreachable, auth-failed,
            script-error, unsupported, unknown, internal.
          example: unreachable
        message:
          type: string
    ApiErrorEnvelope:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ApiError"
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
    Job:
      type: object
      properties:
        id: { type: string }
        device_id: { type: string }
        operation: { type: string, enum: [deploy, logs-sync] }
        status:
          $ref: "#/components/schemas/JobStatus"
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
        last_message: { type: string }
        error:
          $ref: "#/components/schemas/ApiError"
          nullable: true
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorEnvelope"
    InvalidInput:
      description: The request body failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorEnvelope"
    NotFound:
      description: The device or job ID does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorEnvelope"
    DeviceBusy:
      description: The target device already has a queued or running job.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorEnvelope"
    DeviceError:
      description: The underlying devkit_client operation failed (unreachable device, SSH auth failure, or a script error).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorEnvelope"
...